added CardView component to remove dublicated logic
