+++ src/enums/Enums.tsx
@@ -1,0 +1,9 @@
+export enum ViewType {
+    GRID = "grid",
+    LIST = "list"
+}
+
+export enum SortOptions {
+    fromAtoZ,
+    fromZtoA
+}
